add placeholder classes

+++ ewm-service/src/main/java/ru/practicum/ewm/service/service/request/GetEventsPublicRequest.java
@@ -19,5 +19,5 @@ public class GetEventsPublicRequest {
     private boolean onlyAvailable;
     private EventSort sort;
     private long from;
-    private long size;
+    private int size;
 }
